Separate redux connection from search page declaration.
